refactoring, add one more test
- remove private method test.
- rename test files.
- test with invalid column.
